requirements.txt
fastapi
uvicorn

google-generativeai
ddgs
trafilatura
python-dotenv
pydantic